Modifico la función getToken de authorizationService para que devuelva siempre string. Quito aquellos as string que se necesitaban antes para el token. Reviso páginas referentes al login y corrijo a Pablo funciones que ya tenemos definidas.

+++ src/app/pages/login/login-form/login-form.component.ts
@@ -81,8 +81,11 @@ export class LoginFormComponent {
         this.loginError = 'Usuario o contraseña incorrectos';
       }
     } catch (errorResponse) {
-      this.usersService.remove(token);
-      localStorage.removeItem('token');
+      if (token) {
+        this.usersService.remove(token);
+        localStorage.removeItem('token');
+      }
+
       if (errorResponse instanceof HttpErrorResponse && errorResponse.status === 0) {
         this.loginError = constants.generalServerError;
         return;
